Moved implementation from Scream to ScreamEvaluator.
